Avoid duplicate "Latest update" in metadata
Change description is used as "Latest update" but isn't labelled that
way in the publisher. The frontend didn't add this label until now.

This leds to users appending (in a variety of formats) "Latest update:"
to the start of the change description.

The frontend now has a latest update label (to match usage), so we can
strip this out.

Avoids: "Latest update: Latest update - …"