Use static JSON as travel advice publisher related links and breadcrumbs
The breadcrumbs and related links were previously using the passthrough hash
from travel advice publisher, but we are removing them from the publisher,
and the publishing API and the content store.

There was a brief investigation to see if the passthrough hashes could
conform to the dependency resolution style, but these are still
different to how related links are structured in the other apps, so
leaving this as a work around for now until such a time where we can
simplify the structure, or make it conform to how the other apps work.